requirements.txt
pip3 install pycuda --user
h5py==3.6.0
matplotlib==3.5.1
mpi4py==3.1.3
numpy==1.21.5
pymatgen==2022.8.23
PyYAML==6.0
scipy==1.7.3
colorlog
scikit-learn
tqdm
gputil
